requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-cleanup==5.1.0
django-email-verification==0.1.0
django-filter==2.4.0
django-livereload==1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7.1
Pillow==8.1.2
PyJWT==2.0.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
